--- components/H4HStoryCarousel/index.js
@@ -1,3 +1,0 @@
-import H4HStoryCarousel from './H4HStoryCarousel';
-
-export default H4HStoryCarousel;
