+++ components/HeartForHouseComponents/FutureSection/FutureSection.data.js
@@ -1,0 +1,27 @@
+export const CONTENT_BLOCK_DATA = [
+  {
+    title: <u>BOCA RATON</u>,
+    description:
+      '<div style="font-weight: 200;">Christ Fellowship Boca Raton has experienced exponential growth ever since opening its doors. While our original plans included the demolition of our current site to build a 400-seat auditorium, we realized the current growth trajectory would eventually outgrow those plans! As we began to pray, God opened new doors right down the road. <br/><br/>Earlier this year, we purchased the former Macy’s Furniture Store on Glades Road—a 50,000-square-foot building, sitting on 5 acres of land that we intend to transform into a 1,000-seat sanctuary with ample space for ministry for every generation. <div  style="display: inline; font-weight: 600; color: #E63E51;">With a population of nearly 100,000 people, a large majority being unchurched—Boca Raton boasts opportunities to reach people for Jesus!</div> We anticipate this new home will be ready by Christmas 2025, as we transform this store into a place for God to do exceedingly more!</div>',
+    image: '/heart-for-house/future-boca.jpeg',
+    orientation: 'left',
+  },
+  {
+    title: <u>JUPITER</u>,
+    description:
+      '<div style="font-weight: 200;">Last year, we shared the news that we were finally planting permanent roots in the town of Jupiter as we purchased the former Jupiter Fitness Center building located in the heart of Jupiter on Indiantown Road. Our plans include a complete renovation of the 20,000-square-foot space into a 400-seat sanctuary. But while awaiting these renovations and continuing to meet at the Wyndham Grand, God opened a door right next door. <br/><br/>After six years as a portable campus, in the Fall of 2023, Christ Fellowship Jupiter moved into its first-ever permanent location—a move-in ready, former church building located in the same parking lot as our future location! <div style="display: inline; font-weight: 600;  color: #E63E51;">Once the new building is renovated, we will be able to use both spaces for multi-generational ministry, including ample space for kids, students, and classes.</div> We anticipate opening the doors by Easter 2025—transforming this space where people worked out to a place where people can work out their faith and grow in their relationship with Jesus!</div>',
+    image: '/heart-for-house/future-jupiter.jpeg',
+    orientation: 'right',
+  },
+  {
+    title: (
+      <>
+        CFSEU <u>HOUSING</u>
+      </>
+    ),
+    description:
+      '<div style="font-weight: 200;">As a part of our Get There First initiative, we are expanding Southeastern University at Christ Fellowship (CFSEU) in the years to come—with a vision to provide outstanding Christian leadership training through accredited classes for Dual Enrollment, Undergraduate, and Master’s Degrees.<br/><br/> We have a conceptual plan to build a <div style="display: inline; font-weight: 600; color: #E63E51"> multi-unit residence that will house both students and resident interns</div>—opening enrollment for prospective students across the country, and making CFSEU a premier destination in Christian education! As a first step in this large project, this year, we’ll be starting underground work and preparing the land for construction.</div>',
+    image: '/heart-for-house/future-cfseu.png',
+    orientation: 'left',
+  },
+];
