+++ components/HeartForHouseComponents/FutureSection/index.js
@@ -1,0 +1,3 @@
+import FutureSection from './FutureSection';
+
+export default FutureSection;
