+++ components/HeartForHouseComponents/StoryCarousel/StoryCarousel.components.js
@@ -117,33 +117,3 @@ export const TitleSection = () => (
     </Box>
   </Box>
 );
-
-export const CAROUSEL_SLIDE_DATA = [
-  {
-    title: 'CORRADO AND DANTE',
-    content:
-      '“Christ Fellowship got there first in my life by being the rock I grew up on, and the reason my faith has grown to where it is today.” ',
-    author: 'CORRADO',
-    image: '/heart-for-house/carousel/corrado-dante.jpg',
-    url: 'https://issuu.com/christfellowshipchurch/docs/clients_heartforthehouse_print_handouts_7.5x10_dig/23?fr=sOTllNjcyMzY3MDQ',
-    aspectRatio: '16by9',
-  },
-  {
-    title: 'BIANCA AND AIDEN',
-    content:
-      '“After MVMNT Camp, my perspective changed. I looked forward to church again, cracked open my Bible more, joined the Called Retreat, and started attending Celebrate Recovery. Because of Shaun, Jim & Lynann, my family, and everybody who has given to Heart for the House and the vision to Get There First for my generation—I see hope because of the people I’m surrounded by.”',
-    author: 'AIDEN',
-    image: '/heart-for-house/carousel/biance-tejeda.jpg',
-    url: 'https://issuu.com/christfellowshipchurch/docs/clients_heartforthehouse_print_handouts_7.5x10_dig/14?fr=sMDNlZTcyMzY2OTk',
-    aspectRatio: '4by3',
-  },
-  {
-    title: 'THE INVERSO FAMILY',
-    content: `“While Logan’s life ended in a tragic and unexpected way, his journey still continues through our family. Today, you’ll find us serving every Sunday in the place that was always there for us.
-      And our daughter Taylor has found spiritual siblings to champion her calling. We simply can’t get enough of encouraging others, because that’s the encouragement that was given to us.`,
-    author: 'TRACY & JONATHAN',
-    image: '/heart-for-house/carousel/inverso-family.jpg',
-    url: 'https://issuu.com/christfellowshipchurch/docs/clients_heartforthehouse_print_handouts_7.5x10_dig/26?fr=sYzY1ZDcyMzY3MDM',
-    aspectRatio: '4by3',
-  },
-];
