+++ components/HeartForHouseComponents/StoryCarousel/index.js
@@ -1,0 +1,3 @@
+import StoryCarousel from './StoryCarousel';
+
+export default StoryCarousel;
