--- components/HeartForHouseContentBlock/index.js
@@ -1,3 +1,0 @@
-import HeartForHouseContentBlock from './HearForHouseContentBlock';
-
-export default HeartForHouseContentBlock;
